server/requirements.txt
Django==1.5.4


# PostgreSQL database driver. TODO: make the project db-agnostic.
psycopg2==2.4.1

Pillow==2.3.0

# editable maps for forms
django-olwidget==0.47

# used for generating PDF invoices
pisa==3.0.33
reportlab==2.5
html5lib==0.90


# Testing tools
nose==1.1.2
selenium==2.25.0
mock==1.0.0
WebTest==1.3.3
django-webtest==1.5.2


# the latest release of djangosanetesting (0.5.11) doesn't work with the new-
# style DATABASES settings, so we'll just use this until there is a compatible
# version released


# Documentation system
Sphinx==1.1.3


# useful for parsing timestamps from API calls
python-dateutil==2.1


# Command line utils
clint==0.3.1


# DB schema migrations
South==1.0


# some useful utils :-)
pipetools==0.2.0


# Automatic sane defaults for ModelAdmin
#flexipy library for communication with Flexibee
flexipy


#django rest framework
djangorestframework==2.4.4
django-rest-swagger==0.2.9
django-cors-headers==1.0.0

markdown==2.5.1
django-filter==0.8
django-nose==1.3

# Helpdesk
django-markdown-deux==1.0.5
django-bootstrap-form==3.1
simplejson==2.3.3
email-reply-parser==0.3.0
